requirements.txt
blinker==1.8.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.1.1
cycler==0.12.1
easyocr==1.7.2
filelock==3.16.1
flask==3.0.3
fonttools==4.56.0
fsspec==2025.2.0
huggingface-hub==0.29.1
idna==3.10
imageio==2.35.1
importlib-metadata==8.5.0
importlib-resources==6.4.5
itsdangerous==2.2.0
jinja2==3.1.5
kiwisolver==1.4.7
lazy-loader==0.4
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
networkx==3.1
ninja==1.11.1.3
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.1.105
opencv-python==4.11.0.86
packaging==24.2
pandas==2.0.3
pillow==10.4.0
psutil==7.0.0
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pyparsing==3.1.4
pytelegrambotapi==4.23.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pytz==2025.1
PyWavelets==1.4.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-image==0.21.0
scipy==1.10.1
seaborn==0.13.2
shapely==2.0.7
six==1.17.0
sympy==1.13.3
telebot==0.0.5
tifffile==2023.7.10
tokenizers==0.20.3
torch==2.4.1
torchvision==0.19.1
tqdm==4.67.1
transformers==4.46.3
triton==3.0.0
typing-extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.80
ultralytics-thop==2.0.14
urllib3==2.2.3
werkzeug==3.0.6
zipp==3.20.2
